Revert "REVERTME [VPG]: drm/i915: Fix m2 programming for CHT"
This reverts commit efb1b7659c943dc8e21f75522f2284b63943b5e5.
DP m2 values alone required additional shift of 2 bits since
the precalculated values were with the shift.
The pll calculation logic used for hdmi is implemented in
such a way that it considers this shift thus requiring
special handling for DP alone. Once this patch is reverted
and the precalculation is updated both DP and HDMI pll values
will have m2 at the same offset and work as expected.

Change-Id: Ia62da90ee011612640326c7bb72a753b591a5ec5
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-19502
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/468181