cpufreq: intel_pstate: Perf/Power improvement for the load-based algorithm
These changes aim to improve performance and power for the load-based
algorithm by detecting the oscillating workloads (e.g IcyRocks) and
preventing the migrations.

The first change is to improve the accuracy of the pstate decision by
using the average pstate instead of the current p-state. For most of the
cpus, the current pstate (latest requested pstate) does not represent the
real pstate since the cpus are all sharing the same frequencies in a given
module.

The second patch is detecting the use cases such as gaming where the
workload is moving a lot and where we always took the decisions too late.
I'm using a 64-bit bitfield where each bit is set when a high load is
detected. We are increasing the floor of the pstate min proportionnaly to
the number of high loads up to HFM (so the perf and fps increases with a
minimal power impact).

Here is the power and perf comparison versus the 'performance' algorithm:

Benchmark               ∆Perf    ∆Power
SmartBench-Gaming       -0.8%   -31.6%
SmartBench-Productivity -0.9%   -31.6%
Circular Progress Bar     n/a   -65.2%
CandyCrush                n/a   -52.6%
CutTheRope                n/a   -58.7%
AngryBirds                n/a   -47.1%
videoPlayback             n/a   -68.6%
audioPlayback             n/a   -22.8%
IcyRocks-0-0            -0.1%   -58.4%
IcyRocks-20-50          -0.1%   -57.8%
IcyRocks-40-100         -0.3%    -9.0%
IcyRocks-60-150         -1.2%    -1.9%
IcyRocks-80-200          0.2%    -2.3%
IcyRocks-100-250         0.7%    -3.0%
RL Benchmark-Overall    -6.3%   -20.2%

Change-Id: I71e3a14fe2791cc6e0862d9b0af9dd442e279d14
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-15868
Signed-off-by: Philippe Longepe <philippe.longepe@intel.com>
Reviewed-on: https://android.intel.com:443/456778