[EM/PMIC_CCSM]: Fix enumeration issue post data role swap
Sometimes device is not getting enumerated post data role swap. This
is due to pmic ccsm driver not notifying EVENT_VBUS to the otg driver,
since by default is_notify_otg flag is set to false and also after
connecting and disconnecting wall charger, pmic ccsm driver not
notifying to otg as the flag set to flase. Hence enabled the
is_notify_otg flag by default, so by default all the events will be
notified to otg driver and also enabled the flag to notify otg if the
wall charger is removed.

Change-Id: Ic36239c401e6c4a36fadc231fcd1fc33bb1dd2d0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-16390
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Reviewed-on: https://android.intel.com:443/459742