FOR_UPSTREAM [VPG]: drm/i915: Optimize CD clock change for CHT
Current logic that tracks CD clock requirement changes
the clock value whenever a plug in or plug out a new
display. CD clock change during plug in is a must to
support new display if the panel requires it but
the change back to lower CD clock value during unplug
can be avoided. Since the display is bound to be
entering suspend state anytime later we can lower
the cd clock to the required value during suspend.
This also helps in better user experience by avoiding
blanks (disable - enable) of local display.

Change-Id: I5bf52af0af5d49185474d308cb7962b4aca0c3ac
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-27481
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/496010