usb: gadget: Fix mtp_read freeze when switching device config to rndis and then back to mtp
When switching config to rndis,acm,adb, mtp_read should return when
device is offline, instead its stuck on wait_event_interruptible,
waiting for device to be online. Its because mtp device is always
offline when config does not include mtp mode. Because of that process
is put to sleep and usb_mtp driver is not released as expected, this
prevents MtpService from reopening usb_mtp when configuration is
switched from rndis to mtp.

Change-Id: Ia430e4b63f587c3c40fb86808033a7ad45c6dbc0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-12506
Signed-off-by: Robert Dobrowolski <robert.dobrowolski@intel.com>
Reviewed-on: https://android.intel.com:443/460312