iio: ltr501: fixed issue with regmap_bulk_read not returning correct val
it was observed that the regmap_bulk_read call within ltr501_read_als
did not update the sensor read value with every call to it,
The value was read back from the cache itself.
Updated the cache_type to REGCACHE_NONE to read the value directly from
the sensor register.

Change-Id: I9cfd0ae64e8aaf77542fb8f60d8c60f2d4c3e597
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-25780
Signed-off-by: Arun Radhakrishnan <arun.radhakrishnan@intel.com>
Reviewed-on: https://android.intel.com:443/476587
Reviewed-on: https://android.intel.com:443/484462