FOR_UPSTREAM [VPG]: drm/i915: Change sleep to delays for DPCD ops
DPCD operations are time sensitive and using sleep based calls such
as msleep results in some calls taking upto 10ms to be returning.
Hence this patch changes such calls in the dpcd path to
usleep_range. this will guarantee that the required delay is
satisfied and will also ensure that we continue normal
processing immediately after this delay.

Change-Id: I9eede5c10bc30c502271c6f801f529ab44c113e2
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-27481
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/496013