FOR_UPSTREAM [VPG]: drm/i915: use correct pipe for get/put in dp aux
get/put calls for power domain had been hardcoded domain in CHT.
This patch corrects this by using the correct domain based on the
call to intel_dp_aux_ch. Since a call to intel_dp_aux_ch can be
from detect path during which time a pipe may not be associated
with the encoder/connector we reverse calculate using the
other displays enumerated to identify the pipe the request
is made for.

Change-Id: I75b83df01432221ca222a125e3dc1b34fb48ce33
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31683
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/498188
Reviewed-on: https://android.intel.com:443/504758