SQUASHME! [VPG]: drm/i915: fix boundry condition for allow_dpcd
allow_dpcd was set during the port enable itself, this resulted
in a timing issue where a short pulse is received before link
training is complete, since we allow dpcd access during port
enable itself we try to read link status and find that it is
not up. this in turn triggers fake detach and attach. To
avoid this scenario it is better to allow dpcd access
post link training alone.

In addition this patch fixes the boundry condition so
we will exit when we reach the retry count.

This patch has to be squashed into the following patch
Change-Id: Id67e8ae7701c2eed6d04b0fb715a764eac6b9b2f
Author: Thulasimani,Sivakumar <sivakumar.thulasimani@intel.com>
Date:   Tue Nov 3 14:49:29 2015 +0530
FOR_UPSTREAM [VPG]: drm/i915: delay short pulse handling till display
is up

Change-Id: Ic9a364c853b43226407511feee86710e2ebdbadf
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-27481
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/496017