SQUASHME! [VPG]: drm/i915: Always do fake detach-attach for SHPD failure
It was incorrect to assume that forcing full detect when
handling short pulse interrupt failures will recover the
system. It will just end up detecting the panel as still
connected and exit. hence force fake detach and attach
for all short pulse failures to allow the system
to be recovered.

Upfront logic is not required for simulated detach and attach since
it is expected that the lane count will not change through
short pulse. This also helps in avoiding the time spent in
upfront link training.

This patch has to be squashed to the following patch
Author: Thulasimani,Sivakumar <sivakumar.thulasimani@intel.com>
Date:   Thu Sep 24 18:53:36 2015 +0530
FOR_UPSTREAM [VPG]: drm/i915: Lane count change detection

Change-Id: I5fe548fa2be697e0aafb54f04bfbeb1de127bc1e
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31683
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/498192
Reviewed-on: https://android.intel.com:443/504754