UVC: Make UVC_URB a module paramater
On new USB3 based webcams we observe that the number of URBs must to be
increased in order avoid buffer underruns, leading to image tear and
drops.

Since URB mamory is expensive, it is preferrable to have this as a module
parameter rather then just increase the USB_URBS define.

Change-Id: I81284292faad83efd5d00422ad1188f8fecdcd6e
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31683
Signed-off-by: Aviv Greenberg <aviv.d.greenberg@intel.com>
Reviewed-on: https://android.intel.com:443/496205
Reviewed-on: https://android.intel.com:443/492581