XPWR CHARGER: Correcting the CV target voltage setting
Issue description: The driver was incorrectly checking for the
cv values. For example if the max voltage is 4.2V, then driver would
set the target voltage as 4.35V and for max voltage < 4.2V, the
driver was setting the target voltage to 4200mV. As a result, the
battery was never hitting "FULL" status as the charger would never
terminate charging based on ocv.

Solution: Modified the check from less than to less than equal to.
So for values <=4.2V, the target voltage would be 4.2V.

Change-Id: I44c84121cebc859c28f1b04bc4d27ab378989d98
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-29166
Reviewed-on: https://android.intel.com:443/494106
(cherry picked from commit bcf54dd3d99cfc1f3409f23d6f027399d228dc40)
Reviewed-on: https://android.intel.com:443/494596