iio:magn:ak0911: optimizing hrtimer trigger handler
it was observed that we were spending a large chunk of time within
the trigger handler individually reading the data for x,y,z axises
one by one within a loop. Optimizing this by block reading all the
axises at once

According the Datasheet the AK09911_REG_ST2 register read implies
the data read cycle is over and the x,y,z data values can be
refreshed by the sensor. Moving the ST2 register read to as soon
as the data read finishes successfully so that the next read cycle
can progress while the data that has been read can be processed and
pushed to the iio buffers by the driver.

Change-Id: Ib2dff61dfcb7306c4f1b6c480e579e14a2fb5a4c
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-32194
Signed-off-by: Arun Radhakrishnan <arun.radhakrishnan@intel.com>
Reviewed-on: https://android.intel.com:443/506341