input: touchscreen: silead: deactive debug logs set manually
Touchscreen driver is very verbose and as this loading the console
output. The system is loaded with a high priority process starving
out other clients from serving interrupts. System performance is
affected by scheduling the serial output too much. Logs are sent
to the console every time there is an interaction with the device
by touching the screen.

Change-Id: Icb21801911868cc0313f1db162e4a00b8ae57144
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31785
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>
Reviewed-on: https://android.intel.com:443/501512
Reviewed-on: https://android.intel.com:443/504775