Input: touch: silead: Solve touch reading to detect 10 fingers.
The root cause of this issue: i2c_smbus_read_i2c_block_data
function can read up to I2C_SMBUS_BLOCK_MAX (It was defined
 32 bytes as specified in SMBus standard) bytes at a time.
But we use i2c_smbus_read_i2c_block_data function to read
0x80 (counter about finger numbers) and 10 points coordinates.

In fact, A total of 44 bytes are required to read.
So the extra 12 bytes are not normally read, and the final 3
 points of the coordinates can not be detected.

The solution of this issue:
We should use i2c_smbus_read_i2c_block_data function to read
 twice times. Read for the first time 12 bytes, and finally
read the 32 bytes.

Change-Id: I293b2dc1edd1b1c708365cf85ef54adbd8746302
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31785
Signed-off-by: Yu Xia <xiax.yu@intel.com>
Signed-off-by: Ulises Mendez Martinez <ulises.mendez.martinez@intel.com>
Reviewed-on: https://android.intel.com/#/c/501683/
Reviewed-on: https://android.intel.com:443/501683
Reviewed-on: https://android.intel.com:443/504782