sdhci.c: uses TSC instead of jiffies in atomic context.
We hit watchdog panic when sdhci_execute_tuning.

Due to some hardware issue, sdhci int status might not
be available for a long time. The driver used jiffies
to wait for 150 ms while local irq disabled. If the thread
is running on a cpu who is the owner to update jiffies,
the timer irq can’t be processed because the cpu disable
irq. That made the driver wait for more time than needed
and trigger watchdog.

The patch uses TSC instead of jiffies in atomic context.

Change-Id: I90f0fac987d6b4754a0ea7097ac1cf48ad0e9732
Signed-off-by: xiao jin <jin.xiao@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31387
Reviewed-on: https://android.intel.com:443/503999