Fixup for Merge Commit of v3.14.70 and 6262511
This is a fixup for the merge commit of v3.14.70 and the
google common.git android-3.14 branch (commit=6262511).

The branch is created as a basis for doing a merge-commit
of the google common.git android-3.14 branch into an intel
android kernel branch. Future merge-commits can be done
directly with the commit.git project. In order to get this
one right, the fixups are based on a comparison with a
linear quilt representation of the same source code.

Change-Id: I431fed2e2725f44dcab2cd6f90a19db4b4c78a7a
Signed-off-by: Russ Weight <russell.h.weight@intel.com>