DSP should be able to recover after a crash
[Problem]
In certain use cases, DSP crashes, and it can not be recovered. We need to
find a way to have DSP recovered itself than relying on user space to
restart audio pipeline.

[Solution]
Add recovery mechanism to DSP if it times out. Put it to reset state, and
download fw again in this case.

Change-Id: Ica2111c4e45165a98482dad83320a7fd025e2cb2
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-43061
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-42168
Signed-off-by: Ying.Wang <yingwan@lab126.com>
Signed-off-by: Mirche, Dinesh <dinesh.mirche@intel.com>
Reviewed-on: https://android.intel.com:443/570269