Waiting TX messages be sent before close clients in ishtp client driver.
Issue:
In current ishtp client driver, when a client is closed, all tx buffers
will be removed even there are some messages not sent. So, if an application
calls write() to write a message to tx buffer, then calls close() immediately,
the previous message has big chance be removed from tx buffer and can not
be sent to ishfw.

Fix is:
Because there is flow control mechanism between client driver and ishfw for
each message sending, and flow control handled by client driver, so the bese
way is keep waiting if tx buffers are not empty during close() calling,
Until all messages are sent and tx buffers are empty.

Change-Id: I82d4a6a462f428ffa8e03b5e582ae8c1618df5d1
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-35771
Signed-off-by: Even Xu <even.xu@intel.com>
Reviewed-on: https://android.intel.com:443/524445