REVERTME [VPG]: drm/i915: Perfmon pin/unpin oa buffer support
Workaround for an issue related to coreu permission level change.
In latest Android configuration coreu doesn't have SYS_ADMIN capability.

REVERTME: perfmon is due to be replaced in the future, to leverage
the new perf framework currently being upstreamed. When the required
support is available in the i915 and perfmon umd toolset, this patch
will be reverted.

Signed-off-by: Sebastian Szulc <Sebastian.Szulc@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-37923
Change-Id: I56621b166b3cb9d3fd925a850edb639fa6298fb6
Reviewed-on: https://android.intel.com:443/547274