usb/dwc3: wait for EP transfer to complete before disable dwc3 irq
Currently, in "for (epnum = 2; epnum < DWC3_ENDPOINTS_NUM; epnum++)",
it is waiting the DWC3_EP_END_TRANSFER_PENDING flag to be 0 which is
cleared in dwc3_endpoint_interrupt(). However, before that,the dwc3 irq
was disabled which reasult this flag will not be cleared.

This patch moves "for (epnum = 2; epnum < DWC3_ENDPOINTS_NUM; epnum++)"
ahead of "dwc3_gadget_disable_irq(dwc)" to fix this case.

Change-Id: I984835cbd43e37dad2c6bc9cc4281090b7e99253
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-40604
Signed-off-by: Chuansheng Liu <chuansheng.liu@intel.com>
Signed-off-by: Zhen Han <zhen.han@intel.com>
Reviewed-on: https://android.intel.com:443/555057