dwc3: don't set DWC3_EP_END_TRANSFER_PENDING flag if ENDTRANSFER fails
According the dwc3 spec, when issuing an End transfer command, software
must set the CmdIOC bit so that an "Endpoint Command Complete event"
is generated after this transfer command.
In the driver, the DWC3_EP_END_TRANSFER_PENDING is used to
indicate whether certain transfer is pending.

However, if the dwc3_send_gadget_ep_cmd() sent failure,
it should not set the flag DWC3_EP_END_TRANSFER_PENDING
as no transfer is in on-going. This patch fixes this case.

Change-Id: Idb787a29839735e6940885b82f681761edc77e05
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-40604
Signed-off-by: Chuansheng Liu <chuansheng.liu@intel.com>
Signed-off-by: Zhen Han <zhen.han@intel.com>
Reviewed-on: https://android.intel.com:443/555069